requirements.txt
django>=3.1.4
djangorestframework>=3.12.2
pandas>=1.2.2
pymongo>=3.11.3